settings: Fix filter alignment under headings in all settings panels.
Previously, when resizing the window, the filters/buttons and the header
did not fit on the same row, the filters/buttons were pushed onto the
next row but not put directly below the heading. This changes the
settings_panel_list_header and add_default_streams_button_container into
flex containers and properly aligns the items below the heading when
changing window size. Fixes #30353.